requirements.txt
anthropic>=0.49.0
pydantic>=2.0.0
pyyaml>=6.0
genanki>=0.13.0
gtts>=2.5.0
requests>=2.31.0
python-dotenv>=1.0.0
python-telegram-bot>=20.0
exceptiongroup==1.1.3
pytest>=7.4.0,<9.0.0
pytest-mock>=3.12.0
ruff>=0.4.0
